+++ affogato-structure/src/test/java/com/parsuomash/affogato/structure/validation/PasswordTest.kt
@@ -3,7 +3,7 @@ package com.parsuomash.affogato.structure.validation
 import com.google.common.truth.Truth.assertThat
 import org.junit.jupiter.api.Test
 
-internal class PasswordCheckerTest {
+internal class PasswordTest {
   @Test
   fun `check password is valid`() {
     val password = password {